+++ src/java_class.rs
@@ -7,7 +7,7 @@ use crate::cp_info::{parse_cp_info, CpInfo};
 use std::collections::HashMap;
 use std::io::Cursor;
 
-#[derive(Default)]
+#[derive(Default, Debug)]
 pub struct ClassFile {
     /// ClassFile Structure
     /// https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-4.html#jvms-4.1
@@ -100,18 +100,4 @@ fn test_parse_class() {
 
     assert_eq!(result.attributes_count, 1);
     assert_eq!(result.attributes.len(), result.attributes_count as usize);
-}
-
-pub struct Class {
-    pub methods: HashMap<String, MethodInfo>,
-    pub fields: HashMap<String, FieldInfo>,
-}
-
-impl Class {
-    pub(crate) fn createFrom(class_file: ClassFile) -> Class {
-        Class {
-            methods: HashMap::new(),
-            fields: HashMap::new(),
-        }
-    }
 }
